Move shared pytest fixtures to tests.utilities
