+++ iati/core/tests/test_resources.py
@@ -4,7 +4,7 @@ import pytest
 import six
 import iati.core.constants
 import iati.core.resources
-
+from iati.core.tests.utilities import codelist_lengths_by_version, standard_version_optional
 
 class TestResources(object):
     """A container for tests relating to resources."""
