+++ iati/core/tests/test_schemas.py
@@ -6,6 +6,7 @@ import iati.core.default
 import iati.core.exceptions
 import iati.core.schemas
 import iati.core.tests.utilities
+from iati.core.tests.utilities import standard_version_optional
 
 
 class TestSchemas(object):
